pyproject.toml
[tool.poetry]
name = "homework-buster-api"
version = "1.0.0"
description = "Backend API for Homework Buster mobile app"
authors = ["Morgan Chorlton <morganchorlton3@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.13"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
mypy = "^1.11.0"


[tool.poetry.group.api.dependencies]
fastapi = {extras = ["standard"], version = "^0.123.0"}
mangum = "^0.19.0"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"
pydantic = "^2.12.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"


[tool.ruff]
line-length = 100
target-version = "py311"

